Find sum of all multiples of 3/5 below given integer
This program sums all multiples of 3/5 for the range of given integer. 
the program contains a comment which explains what is the aim of this file.
It contains a main function which gets run when the file is executed.
The function sum_up_3_or_5_multiples calculates the sum of integers that are multiples of 3/5 within the  upper range.