Implement python generators
Generators have similar functionalities as lists. But they are used in
order to save memory and time of execution.
When a generator is implemented, the complete value of the enumerator to
be returned is not caluclated. It is calculated on the fly, when it is
called.
If you try to print a generator, you will just be able to see a class
called generator.
If you want to see the items of the generator, loop through to print
them. When you loop, the values are calculated on the fly.