Implment insert, search, traversal in BST
A Binary Search tree is a special form of tree with strictly 0 or 2
nodes as children for each node.
It can be left or right node such that
left child of node < node < right child of the node.
Eg. (left=5) < (parent=10) < (right=15)
    10
    /\
   5  15
The code also contain three traversal.
BST mostly implements in-order traversal.

The insert and find for the BST is O(h), where h = height of the tree.
It is usually O(log n)
This is true for a balanced tree.

For an unbalanced tree, it could be iefficient and behave like a list
with O(n) for insert and find.